Refactor CoseKey and CoseKeyPrivate to use CoseEcCurve enum instead of ECCurveType
Refactor the CoseKey and CoseKeyPrivate structs in the CoseKey.swift file to use the CoseEcCurve enum instead of the ECCurveType enum. This change ensures consistency and clarity in the codebase.

Refactor DeviceEngagement to use CoseEcCurve enum in initializer

Refactor the DeviceEngagement struct in the DeviceEngagement.swift file to use the CoseEcCurve enum in the initializer instead of the ECCurveType enum. This change improves code readability and maintainability.

Update CoseEcCurve to use UInt64 raw value type

Update the CoseEcCurve enum in the CoseEcCurve.swift file to use UInt64 as the raw value type instead of Int. This change aligns with the COSE registry and ensures compatibility with iOS natively supported curves.

Refactor KeyOptions.swift to use requireApplicationPassword instead of requirePassword

Refactor the KeyOptions.swift file to use the requireApplicationPassword property instead of the requirePassword property in the KeyAccessControl struct. This change improves clarity and consistency in the codebase.